finish doctorProfile & some touches & khaled work

+++ JUPReact/src/Components/Header/Navbar2.jsx
@@ -1,8 +1,13 @@
+import React, { useState, useContext } from "react";
 import Container from "react-bootstrap/Container";
 import Nav from "react-bootstrap/Nav";
 import Navbar from "react-bootstrap/Navbar";
-import { NavDropdown } from "react-bootstrap";
+import { NavDropdown, Button } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
+import { faSearch } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { NutFill, Search } from "react-bootstrap-icons";
+import axiosClient from '../../axios';
 
 //css
 import "./Navbar2.css";
@@ -10,22 +15,47 @@ import "./Navbar2.css";
 //images
 import Logo from "./sandralogo.png";
 import icon from "./usericon.png";
-import { NutFill } from "react-bootstrap-icons";
 
+//components
+import SearchScreen from "../Search/Search";
+
+//providers
+import { SearchContext } from "../../Providers/SearchProvider";
 function Navbarr() {
+    const { query, setQuery, click, setClick } = useContext(SearchContext);
     const navigate = useNavigate();
+    const [showModal, setShowModal] = useState(false);
 
     function Logout() {
-        localStorage.clear();
-        navigate("/login");
+        const userType = localStorage.getItem("user-type");
+        const userId = localStorage.getItem("user-id");
+
+        // Make a POST request to the logout API
+        axiosClient
+            .post(`/logout/${userType}/${userId}`)
+            .then((response) => {
+                console.log(response.data.message); // Success message from the backend
+                localStorage.clear();
+                navigate("/login");
+            })
+            .catch((error) => {
+                console.error(error);
+                // Handle any errors that occur during the API call
+            });
     }
     let id = localStorage.getItem("user-id");
     let Type = localStorage.getItem("user-type");
 
     let user = localStorage.getItem("user-name");
     console.log(user);
+
+    const handleClickSearch = () => {
+        setClick(true);
+        console.log(click);
+    };
+
     return (
-        <Container>
+        <Container className="Container">
             <Navbar className="Nav" expand="md" fixed="top">
                 <Container fluid>
                     <Navbar.Brand href="/user">
@@ -45,26 +75,60 @@ function Navbarr() {
                     <Navbar.Toggle aria-controls="navbarNav" />
                     <Navbar.Collapse id="navbarNav">
                         <Nav className="me-auto">
+                            <Nav.Link href="/user" className="l">
+                                Home
+                            </Nav.Link>
+
+                            <Nav.Link href="/doctor" className="l">
+                                Doctors
+                            </Nav.Link>
+                            <Nav.Link href="/articles" className="l">
+                                Articles
+                            </Nav.Link>
+
                             {localStorage.getItem("user-info") ? (
                                 <>
-                                    <Nav.Link href="/user" className="l">
-                                        Home
-                                    </Nav.Link>
                                     <Nav.Link href="#chatbot" className="l">
                                         Talk to Sandra
-                                    </Nav.Link>
-                                    <Nav.Link href="/doctor" className="l">
-                                        Doctors
-                                    </Nav.Link>
-                                    <Nav.Link href="/articles" className="l">
-                                        Articles
                                     </Nav.Link>
                                     <Nav.Link href="/chats" className="l">
                                         Chats
                                     </Nav.Link>
+                                    {localStorage.getItem("user-type") !="patient"&&localStorage.getItem("user-type")!="admin"? (
+                                        <Nav.Link href="/verfiy" className="l">
+                                            verfiy
+                                        </Nav.Link>
+                                    ) : null}
+                                    <Nav.Link className="l">
+                                        <div className="s">
+                                            <button
+                                                className="button"
+                                                onClick={handleClickSearch}
+                                            >
+                                                <FontAwesomeIcon
+                                                    className="star"
+                                                    icon={faSearch}
+                                                ></FontAwesomeIcon>
+                                            </button>
+                                        </div>
+                                    </Nav.Link>
+                                    
                                 </>
                             ) : (
                                 <>
+                                    <Nav.Link className="l">
+                                        <div className="s">
+                                            <button
+                                                className="button"
+                                                onClick={handleClickSearch}
+                                            >
+                                                <FontAwesomeIcon
+                                                    className="star"
+                                                    icon={faSearch}
+                                                ></FontAwesomeIcon>
+                                            </button>
+                                        </div>
+                                    </Nav.Link>
                                     <Nav.Link href="/signup" className="l">
                                         Signup
                                     </Nav.Link>
@@ -74,6 +138,7 @@ function Navbarr() {
                                     </Nav.Link>
                                 </>
                             )}
+
                             <div className="info">
                                 {localStorage.getItem("user-info") ? (
                                     <>
@@ -102,7 +167,7 @@ function Navbarr() {
                                             >
                                                 Settings
                                             </NavDropdown.Item>
-                                            
+
                                             <NavDropdown.Item
                                                 onClick={Logout}
                                                 className="o"
@@ -117,6 +182,7 @@ function Navbarr() {
                     </Navbar.Collapse>
                 </Container>
             </Navbar>
+            {showModal && <SearchScreen onClose={handleCloseModal} id={id} />}
         </Container>
     );
 }
